refactor: move remote code to separate file
